+++ share/uname.go
@@ -1,48 +1,12 @@
 package chshare
 
 import (
-	"bytes"
-	"context"
 	"os/exec"
 	"runtime"
 	"strings"
-	"time"
 
 	"github.com/shirou/gopsutil/host"
 )
-
-type Invoker interface {
-	Command(string, ...string) ([]byte, error)
-	CommandWithContext(context.Context, string, ...string) ([]byte, error)
-}
-
-type Invoke struct{}
-
-func (i Invoke) Command(name string, arg ...string) ([]byte, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-	return i.CommandWithContext(ctx, name, arg...)
-}
-
-func (i Invoke) CommandWithContext(ctx context.Context, name string, arg ...string) ([]byte, error) {
-	cmd := exec.CommandContext(ctx, name, arg...)
-
-	var buf bytes.Buffer
-	cmd.Stdout = &buf
-	cmd.Stderr = &buf
-
-	if err := cmd.Start(); err != nil {
-		return buf.Bytes(), err
-	}
-
-	if err := cmd.Wait(); err != nil {
-		return buf.Bytes(), err
-	}
-
-	return buf.Bytes(), nil
-}
-
-var invoke Invoker = Invoke{}
 
 func Uname() (string, error) {
 	if runtime.GOOS == "windows" {
